Music models work

+++ tests/HartsyInference.Interactive.Tests/InteractiveSessionTests.cs
@@ -46,7 +46,7 @@ public class InteractiveSessionTests
         // The compute thread runs free (repeat-last on underrun) and the bounded output queue
         // drop-oldests under load, so the exact set [0,1,2] is not guaranteed. The real invariant this
         // test pins is ORDER: frames arrive strictly increasing and begin at frame 0.
-        List<long> indices = frames.Select(f => f.Index).ToList();
+        List<int> indices = frames.Select(f => f.Index).ToList();
         Assert.Equal(0, indices[0]);
         for (int i = 1; i < indices.Count; i++)
             Assert.True(indices[i] > indices[i - 1], $"frame order regressed: [{string.Join(", ", indices)}]");
